+++ src/pages/settings/invoice-design/pages/template-designs/pages/edit/components/Body.tsx
@@ -13,25 +13,38 @@ import { payloadAtom } from '../Edit';
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useDebounce } from 'react-use';
-import { Card } from '$app/components/cards';
+import { Card, Element } from '$app/components/cards';
 import Editor from '@monaco-editor/react';
 import { useDesignUtilities } from '../common/hooks';
+import { ValidationBag } from '$app/common/interfaces/validation-bag';
+import { InputField } from '$app/components/forms';
+import { Divider } from '$app/components/cards/Divider';
 
-export function Body() {
+interface Props {
+    errors: ValidationBag | undefined;
+}
+
+export function Body(props: Props) {
     const [payload] = useAtom(payloadAtom);
     const [value, setValue] = useState(payload.design?.design.body);
 
     const { t } = useTranslation();
     const { handleBlockChange } = useDesignUtilities();
+    const { errors } = props;
+    const { handlePropertyChange } = useDesignUtilities();
 
     useDebounce(() => value && handleBlockChange('body', value), 1000, [value]);
 
-    console.log(value);
-    console.log(payload);
-    console.log(payload.design);
-    
     return (
-        <Card title={t('template')} padding="small" collapsed={false}>
+        <Card title={t('template')} padding="small">
+            <Element leftSide={t('name')}>
+                <InputField
+                    value={payload.design?.name}
+                    onValueChange={(value) => handlePropertyChange('name', value)}
+                    errorMessage={errors?.errors.name}
+                />
+            </Element>
+            <Divider />
             <Editor
                 height="100rem"
                 defaultLanguage="html"
